+++ dse_do_utils/scenariomanager.py
@@ -14,10 +14,11 @@ import tempfile
 
 import docplex
 import pandas as pd
-from typing import Sequence, List, Dict, Tuple, Optional
+from typing import Sequence, List, Dict, Tuple, Optional, Union
 
 #  Typing aliases
 from dse_do_utils.utilities import convert_size
+from pathlib import Path
 
 Inputs = Dict[str, pd.DataFrame]
 Outputs = Dict[str, pd.DataFrame]
@@ -82,7 +83,7 @@ class ScenarioManager(object):
     """
 
     def __init__(self, model_name: Optional[str] = None, scenario_name: Optional[str] = None,
-                 local_root: Optional[str] = None, project_id: Optional[str] = None, project_access_token: Optional[str] = None, project=None,
+                 local_root: Optional[Union[str, Path]] = None, project_id: Optional[str] = None, project_access_token: Optional[str] = None, project=None,
                  template_scenario_name: Optional[str] = None, platform: Optional[Platform] = None,
                  inputs: Inputs = None, outputs: Outputs = None,
                  local_relative_data_path: str = 'assets/data_asset', data_directory: str = None):